- New RegistrationController - New Spring Security - DB User reg

+++ src/main/java/com/example/sweater/domain/Message.java
@@ -14,7 +14,7 @@ import javax.persistence.Id;
 public class Message {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id; //ідентифікатор для того щоб розрізняти записи в одній таблиці
+    private Long id; //ідентифікатор для того щоб розрізняти записи в одній таблиці
 
     private String text;
     private String tag;
@@ -28,7 +28,7 @@ public class Message {
         this.tag = tag;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -40,7 +40,7 @@ public class Message {
         this.tag = tag;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
